+++ .techtrain/manifests/railway.json
@@ -30,6 +30,7 @@
     "27": "station27.json",
     "28": "station28.json",
     "29": "station29.json",
-    "30": "station30.json"
+    "30": "station30.json",
+    "31": "station31.json"
   }
 }
