+++ .techtrain/manifests/station15.json
@@ -1,16 +1,13 @@
 {
   "name": "Station 15",
   "id": 15,
-  "prepare": [
-    {
-      "command": "yarn lite-server",
-      "background": true
-    }
-  ],
+  "prepare": [],
   "tests": [
     {
-      "type": "cypress",
-      "specFile": "./cypress/integration/station15.spec.ts"
+      "type": "command",
+      "title": "ローカルサーバでhtmlをブラウザから確認（自己申告）",
+      "command": "echo",
+      "args": []
     }
   ]
 }
