--- .techtrain/manifests/station27.json
@@ -1,16 +1,0 @@
-{
-  "name": "Station 27",
-  "id": 27,
-  "prepare": [
-    {
-      "command": "yarn lite-server",
-      "background": true
-    }
-  ],
-  "tests": [
-    {
-      "type": "cypress",
-      "specFile": "./cypress/integration/station27.spec.ts"
-    }
-  ]
-}
