+++ .techtrain/manifests/station30.json
@@ -1,13 +1,16 @@
 {
   "name": "Station 30",
   "id": 30,
-  "prepare": [],
+  "prepare": [
+    {
+      "command": "yarn lite-server",
+      "background": true
+    }
+  ],
   "tests": [
     {
-      "type": "command",
-      "title": "Node.jsのwebサーバの実行を確認（自己申告）",
-      "command": "echo",
-      "args": []
+      "type": "cypress",
+      "specFile": "./cypress/integration/station30.spec.ts"
     }
   ]
 }
