+++ .techtrain/manifests/station31.json
@@ -1,0 +1,13 @@
+{
+  "name": "Station 31",
+  "id": 31,
+  "prepare": [],
+  "tests": [
+    {
+      "type": "command",
+      "title": "Node.jsのwebサーバの実行を確認（自己申告）",
+      "command": "echo",
+      "args": []
+    }
+  ]
+}
