+++ cypress/integration/station19.spec.ts
@@ -1,22 +1,24 @@
 /**
- * # 「SVGについて知ろう！」- SVG
+ * # 「CSSアニメーションを作ろう！」 - CSS Animation
  */
 
-import { compareColor } from '../utils/compareColor'
-
 describe('Station19', () => {
-  beforeEach('Station19', () => {
+  beforeEach(() => {
     cy.visit('/station19.html')
   })
 
-  it('svgのfillに#008000が適用されている', () => {
-    cy.get('svg').then((svg) => {
-      cy.get('rect').then((rect) => {
-        expect(
-          compareColor(svg.css('fill'), '#008000') ||
-            compareColor(rect.css('fill'), '#008000')
-        ).to.be.true
-      })
+  it('開いたときにアニメーションが走る', () => {
+    cy.get('#animation').then((animation) => {
+      const duration = Number(animation.css('animation-duration').slice(0, -1))
+      expect(duration).to.be.greaterThan(0)
+      expect(animation.css('animation-name')).not.to.be.eq('none')
+    })
+  })
+
+  it('アニメーションが終わった後も要素が表示されている', () => {
+    cy.get('#animation').then((animation) => {
+      expect(animation.css('animation-fill-mode')).not.to.be.eq('none')
+      expect(animation.css('animation-fill-mode')).not.to.be.eq('backwards')
     })
   })
 })
