+++ cypress/integration/station21.spec.ts
@@ -1,20 +1,31 @@
 /**
- * # 「DevToolsのNetworkパネルを使えるようになろう！」 - DevTools Network panel
+ * # 「バリデーションをしよう！」 - Validation check
  */
 
 describe('Station21', () => {
   beforeEach(() => {
-    cy.clock()
     cy.visit('/station21.html')
   })
 
-  it('<p>にレスポンス内容をコピペできている', () => {
-    cy.get('p').then((p) => {
-      const responseJson: { message: string; status: string } = JSON.parse(
-        p[0].innerHTML
-      )
-      expect(responseJson.message).to.be.exist
-      expect(responseJson.status).to.be.exist
-    })
+  it('メールアドレスの入力欄のtypeがemail,必須になっている', () => {
+    cy.get('form')
+      .find('input#email')
+      .should('have.attr', 'type', 'email')
+      .and('have.attr', 'required')
+  })
+
+  it('パスワードの入力欄のtypeがpassword, 必須, 6文字以上になっている', () => {
+    cy.get('form')
+      .find('input#password')
+      .should('have.attr', 'type', 'password')
+      .and('have.attr', 'minlength', '6')
+      .and('have.attr', 'required')
+  })
+
+  it('ニックネームの入力欄が20文字以下になっている', () => {
+    cy.get('form')
+      .find('input#nickname')
+      .should('have.attr', 'type', 'text')
+      .and('have.attr', 'maxlength', '20')
   })
 })
