+++ cypress/integration/station24.spec.ts
@@ -1,5 +1,5 @@
 /**
- * # 「アロー関数を使えるようになろう！」 - Arrow Function Expressions
+ * # 「モジュールのインポート/エクスポートを使えるようになろう！」 - Import/Export Statement
  */
 
 describe('Station24', () => {
@@ -7,9 +7,9 @@ describe('Station24', () => {
     cy.visit('/station24.html')
   })
 
-  it('Arrowの文言が表示される', () => {
+  it('クリアの文言が表示される', () => {
     cy.get('#result').then((result) => {
-      expect(result[0].innerHTML).eq('global data')
+      expect(result[0].innerHTML).eq('Station24をクリア!')
     })
   })
 })
