+++ src/assets/css/station14.css
@@ -1,0 +1,5 @@
+@charset "utf-8";
+
+.center {
+  text-align: center;
+}
