+++ src/scripts/station22-common.js
@@ -1,3 +1,6 @@
-var result = getData();
-var elem = document.getElementById("result");
-elem.innerHTML = result;
+function fetchApi() {
+  const url = "https://dog.ceo/api/breeds/image/random";
+  fetch(url)
+    .then(() => console.log("fetch completed"))
+    .catch(() => console.error("fetch failed"));
+}
