--- src/scripts/station23.js
@@ -1,3 +1,0 @@
-function getData() {
-    return "Station23をクリア!"
-}
