+++ src/scripts/station25-common.js
@@ -1,12 +1,3 @@
-async function main() {
-    const results = await getData();
-    const elem = document.getElementById("result");
-    results.forEach(result => {
-        let li = document.createElement('li');
-        let text = document.createTextNode(result.full_name);
-        li.appendChild(text);
-        elem.appendChild(li);
-    })
-}
-
-main()
+var result = getData();
+var elem = document.getElementById("result");
+elem.innerHTML = result;
