+++ src/scripts/station27-common.js
@@ -1,2 +1,12 @@
-const check = document.getElementById("check");
-check.addEventListener("change", changeBackgroundColor);
+async function main() {
+    const results = await getData();
+    const elem = document.getElementById("result");
+    results.forEach(result => {
+        let li = document.createElement('li');
+        let text = document.createTextNode(result.full_name);
+        li.appendChild(text);
+        elem.appendChild(li);
+    })
+}
+
+main()
