+++ src/scripts/station28-common.js
@@ -1,0 +1,2 @@
+const check = document.getElementById("check");
+check.addEventListener("change", changeBackgroundColor);
