+++ src/scripts/station28.js
@@ -1,0 +1,4 @@
+function changeBackgroundColor(e) {
+  // ここに背景色を変える処理を書く。
+
+}
