--- src/scripts/station29.js
@@ -1,3 +1,0 @@
-function displayList() {
-    // ここに処理を書く
-}
