+++ src/scripts/station30.js
@@ -1,0 +1,3 @@
+function displayList() {
+    // ここに処理を書く
+}
