--- src/station12.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./assets/css/station12.css" type="text/css" />
-    <title>Station 12</title>
-  </head>
-  <body>
-    <div id="red"></div>
-    <div id="blue"></div>
-  </body>
-</html>
