+++ src/station13.html
@@ -4,13 +4,11 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./assets/css/station13.css" type="text/css">
+    <link rel="stylesheet" href="./assets/css/station13.css" type="text/css" />
     <title>Station 13</title>
   </head>
   <body>
-    <p>ブロック要素 pタグ内のテキストだよ<p>
-    <a>インライン要素 aタグ内のテキストだよ</a>
-    <div>ブロック要素 divタグ内のテキストだよ</div>
-    <span>インライン要素 spanタグ内のテキストだよ</span>
+    <div id="red"></div>
+    <div id="blue"></div>
   </body>
 </html>
