+++ src/station14.html
@@ -4,29 +4,13 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./assets/css/station14.css" type="text/css">
     <title>Station 14</title>
   </head>
   <body>
-    <h1>Station 14</h1>
-    <article>
-      <h2>ローカルサーバで確認しよう！</h2>
-      <p>VSCodeのプラグイン<a href="https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer">「Live Server」</a>を使って自分のPC内でwebサーバを立ち上げて<a href="http://localhost:5500">http://localhost:5500</a>にアクセスしてみましょう！</p>
-      <ul>
-        <li>LiveServer:<a href="https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer">https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer</a></li>
-      </ul>
-      <p>こちら↑のURLからVSCode拡張をインストールします。</p>
-      <section>
-        <h3>【インストール手順】</h3>
-        <ol>
-          <li>urlにブラウザでアクセスしてinstallボタンをクリック。</li>
-          <li>「Visual Studio Code is required to install this extension.」というポップアップが出てきた場合はContinueをクリック。(出ない場合もある)</li>
-          <li>「Visual Studio Code を開きますか？」というアラートがでるので開くボタンをクリック。</li>
-          <li>拡張機能: Live Serverのタブが表示されるのでインストールボタンをクリック。</li>
-          <li>VSCodeの右下に「Go Live」の表示が追加されていればokです！</li>
-        </ol>
-        <p>サーバを立ち上げるときはVSCodeの右下の「Go Live」をクリックするとブラウザが開かれるので表示したいファイルを選択してください。</p>
-        <p><img src="./assets/image/live_server_go_live.jpg" alt="VSCode 「Go Live」ボタン" /></p>
-      </section>
-    </article>
+    <p>ブロック要素 pタグ内のテキストだよ<p>
+    <a>インライン要素 aタグ内のテキストだよ</a>
+    <div>ブロック要素 divタグ内のテキストだよ</div>
+    <span>インライン要素 spanタグ内のテキストだよ</span>
   </body>
 </html>
