+++ src/station15.html
@@ -4,20 +4,29 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./assets/css/station15.css" type="text/css" />
     <title>Station 15</title>
   </head>
   <body>
-    <h1>とっても賢い人が書いてるブログ</h1>
-    <h2>成績爆上がり学習法３選！</h2>
-    <p>
-      みなさん学校の成績が伸び悩んで困っていませんか？そんなあなたに学習法を３選紹介したいと思います！！
-    </p>
-    <h3>授業中寝ない</h3>
-    <p>居眠りしないでちゃんと授業を聞きましょう。夜ちゃんと寝ましょう。</p>
-    <h3>復習をしよう</h3>
-    <p>家に帰ったらその日習ったことを書きだしましょう。</p>
-    <h3>朝ごはん食べましょう</h3>
-    <p>朝ごはんちゃんと食べないと頭は働かないのでちゃんと食べましょう。</p>
+    <h1>Station 15</h1>
+    <article>
+      <h2>ローカルサーバで確認しよう！</h2>
+      <p>VSCodeのプラグイン<a href="https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer">「Live Server」</a>を使って自分のPC内でwebサーバを立ち上げて<a href="http://localhost:5500">http://localhost:5500</a>にアクセスしてみましょう！</p>
+      <ul>
+        <li>LiveServer:<a href="https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer">https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer</a></li>
+      </ul>
+      <p>こちら↑のURLからVSCode拡張をインストールします。</p>
+      <section>
+        <h3>【インストール手順】</h3>
+        <ol>
+          <li>urlにブラウザでアクセスしてinstallボタンをクリック。</li>
+          <li>「Visual Studio Code is required to install this extension.」というポップアップが出てきた場合はContinueをクリック。(出ない場合もある)</li>
+          <li>「Visual Studio Code を開きますか？」というアラートがでるので開くボタンをクリック。</li>
+          <li>拡張機能: Live Serverのタブが表示されるのでインストールボタンをクリック。</li>
+          <li>VSCodeの右下に「Go Live」の表示が追加されていればokです！</li>
+        </ol>
+        <p>サーバを立ち上げるときはVSCodeの右下の「Go Live」をクリックするとブラウザが開かれるので表示したいファイルを選択してください。</p>
+        <p><img src="./assets/image/live_server_go_live.jpg" alt="VSCode 「Go Live」ボタン" /></p>
+      </section>
+    </article>
   </body>
 </html>
