+++ src/station16.html
@@ -4,10 +4,20 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Station 16</title>
     <link rel="stylesheet" href="./assets/css/station16.css" type="text/css" />
+    <title>Station 16</title>
   </head>
   <body>
-    <p><a href="">このテキストの色を変化させてね</a></p>
+    <h1>とっても賢い人が書いてるブログ</h1>
+    <h2>成績爆上がり学習法３選！</h2>
+    <p>
+      みなさん学校の成績が伸び悩んで困っていませんか？そんなあなたに学習法を３選紹介したいと思います！！
+    </p>
+    <h3>授業中寝ない</h3>
+    <p>居眠りしないでちゃんと授業を聞きましょう。夜ちゃんと寝ましょう。</p>
+    <h3>復習をしよう</h3>
+    <p>家に帰ったらその日習ったことを書きだしましょう。</p>
+    <h3>朝ごはん食べましょう</h3>
+    <p>朝ごはんちゃんと食べないと頭は働かないのでちゃんと食べましょう。</p>
   </body>
 </html>
