+++ src/station17.html
@@ -4,17 +4,10 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./assets/css/station17.css" type="text/css" />
     <title>Station 17</title>
+    <link rel="stylesheet" href="./assets/css/station17.css" type="text/css" />
   </head>
   <body>
-    <div id="container">
-      <header>ヘッダー</header>
-      <div>アイテム1</div>
-      <div>アイテム2</div>
-      <div>アイテム3</div>
-      <div>アイテム4</div>
-      <footer>フッター</footer>
-    </div>
+    <p><a href="">このテキストの色を変化させてね</a></p>
   </body>
 </html>
