+++ src/station18.html
@@ -8,8 +8,13 @@
     <title>Station 18</title>
   </head>
   <body>
-    <div id="animation">
-      <p>フェードイン</p>
+    <div id="container">
+      <header>ヘッダー</header>
+      <div>アイテム1</div>
+      <div>アイテム2</div>
+      <div>アイテム3</div>
+      <div>アイテム4</div>
+      <footer>フッター</footer>
     </div>
   </body>
 </html>
