+++ src/station20.html
@@ -4,12 +4,12 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./assets/css/station20.css" type="text/css" />
+    <title>Station 20</title>
   </head>
   <body>
-    <main>
-      <form action="" method="post">
-        <button type="submit">送信</button>
-      </form>
-    </main>
+    <svg viewBox="0, 0, 500, 500" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0" y="0" width="100px" height="100px" />
+    </svg>
   </body>
 </html>
