+++ src/station21.html
@@ -7,11 +7,10 @@
     <title>Station 21</title>
   </head>
   <body>
-    <!-- この<p>の中にレスポンスをコピペする -->
-    <p>
-
-    </p>
-    <button onclick="fetchApi()">リクエスト</button>
+    <main>
+      <form action="" method="post">
+        <button type="submit">送信</button>
+      </form>
+    </main>
   </body>
-  <script type="text/javascript" src="./scripts/station21-common.js"></script>
 </html>
