+++ src/station22.html
@@ -5,24 +5,13 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Station 22</title>
-    <script type="text/javascript">
-        function getData() {
-            const argumentArray = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n'];
-            return
-        }
-
-        // ↓こちらの関数は触らないこと
-        function hogehoge(s1, s2, s3, s4, s5, s6, s7, s8, s9, s10, s11, s12, s13, s14) {
-          return s1 + s2 + s3 + s4 + s5 + s6 + s7 + s8 + s9 + s10 + s11 + s12 + s13 + s14
-        }
-    </script>
   </head>
   <body>
-    <div class="card-container">
-      <div class="card">
-        <p id="result"></p>
-      </div>
-    </div>
-    <script type="text/javascript" src="./scripts/station22-common.js"></script>
+    <!-- この<p>の中にレスポンスをコピペする -->
+    <p>
+
+    </p>
+    <button onclick="fetchApi()">リクエスト</button>
   </body>
+  <script type="text/javascript" src="./scripts/station22-common.js"></script>
 </html>
