+++ src/station23.html
@@ -5,6 +5,17 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Station 23</title>
+    <script type="text/javascript">
+        function getData() {
+            const argumentArray = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n'];
+            return
+        }
+
+        // ↓こちらの関数は触らないこと
+        function hogehoge(s1, s2, s3, s4, s5, s6, s7, s8, s9, s10, s11, s12, s13, s14) {
+          return s1 + s2 + s3 + s4 + s5 + s6 + s7 + s8 + s9 + s10 + s11 + s12 + s13 + s14
+        }
+    </script>
   </head>
   <body>
     <div class="card-container">
@@ -12,6 +23,6 @@
         <p id="result"></p>
       </div>
     </div>
+    <script type="text/javascript" src="./scripts/station23-common.js"></script>
   </body>
-  <script type="text/javascript" src="./scripts/station23-common.js"></script>
 </html>
