+++ src/station24.html
@@ -5,7 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Station 24</title>
-    <script type="text/javascript" src="./scripts/station24.js"></script>
   </head>
   <body>
     <div class="card-container">
