+++ src/station25.html
@@ -10,7 +10,7 @@
   <body>
     <div class="card-container">
       <div class="card">
-        <ul id="result"></ul>
+        <p id="result"></p>
       </div>
     </div>
   </body>
