+++ src/station26.html
@@ -4,14 +4,13 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Station 15</title>
+    <title>Station 26</title>
     <script type="text/javascript" src="./scripts/station26.js"></script>
   </head>
   <body>
     <div class="card-container">
       <div class="card">
-          <ul id="result">
-          </ul>
+        <ul id="result"></ul>
       </div>
     </div>
   </body>
