+++ src/station27.html
@@ -5,17 +5,15 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Station 27</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/JNKKKK/MoreToggles.css@0.2.1/output/moretoggles.min.css">
     <script type="text/javascript" src="./scripts/station27.js"></script>
   </head>
   <body>
-    <main>
-      <p id="text">背景色が変わります。</p>
-      <div class="mt-ios">
-        <input id="check" type="checkbox" />
-        <label for="check"></label>
+    <div class="card-container">
+      <div class="card">
+          <ul id="result">
+          </ul>
       </div>
-    </main>
-    <script type="text/javascript" src="./scripts/station27-common.js"></script>
+    </div>
   </body>
+  <script type="text/javascript" src="./scripts/station27-common.js"></script>
 </html>
