+++ src/station28.html
@@ -5,12 +5,17 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Station 28</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/JNKKKK/MoreToggles.css@0.2.1/output/moretoggles.min.css">
     <script type="text/javascript" src="./scripts/station28.js"></script>
   </head>
   <body>
-    <p>アラートを表示させてください。</p>
-    <div>
-      <button>アラートを表示</button>
-    </div>
+    <main>
+      <p id="text">背景色が変わります。</p>
+      <div class="mt-ios">
+        <input id="check" type="checkbox" />
+        <label for="check"></label>
+      </div>
+    </main>
+    <script type="text/javascript" src="./scripts/station28-common.js"></script>
   </body>
 </html>
