+++ src/station29.html
@@ -8,11 +8,9 @@
     <script type="text/javascript" src="./scripts/station29.js"></script>
   </head>
   <body>
-    <div id="fruits">
-      <p>リンゴ</p>
-      <p>イチゴ</p>
-      <p>ブドウ</p>
+    <p>アラートを表示させてください。</p>
+    <div>
+      <button>アラートを表示</button>
     </div>
-    <button onclick="displayList()">リスト表示</button>
   </body>
 </html>
