+++ src/station30.html
@@ -4,12 +4,15 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./assets/css/station19.css" type="text/css" />
-    <title>Station 19</title>
+    <title>Station 30</title>
+    <script type="text/javascript" src="./scripts/station30.js"></script>
   </head>
   <body>
-    <div id="animation">
-      <p>フェードイン</p>
+    <div id="fruits">
+      <p>リンゴ</p>
+      <p>イチゴ</p>
+      <p>ブドウ</p>
     </div>
+    <button onclick="displayList()">リスト表示</button>
   </body>
 </html>
